https-client: look for connection in the app data bag first
If relating with a clustered LXD, the connection info will be in
the app data bag in which case it only makes sense to connect once,
instead of once to each of the cluster members.

Since standalone LXD are also supported, a second lookup needs to be
done in the unit data bag.

Since the https-client is now aware if it relates with a clustered LXD
or not, this knowledge needs to be forgotten when the relation is broken.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>